backened/requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
Djangorestframework==3.12.4
django-cors-headers==3.7.0
Pillow==8.3.0
djangorestframework-simplejwt==4.7.2